Documents now also support custom fields Moved common functions of documents and work items to common base classes
